+++ src/lib/geogram/NL/nl_superlu.c
@@ -575,7 +575,7 @@ static void dgssv_factorize_only(
      *   permc_spec = COLAMD:   approximate minimum degree column ordering
      *   permc_spec = MY_PERMC: the ordering already supplied in perm_c[]
      */
-    permc_spec = options->ColPerm;
+    permc_spec = (int)(options->ColPerm);
     if ( permc_spec != MY_PERMC && options->Fact == DOFACT )
 	SuperLU()->get_perm_c(permc_spec, AA, perm_c);
